+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/UserDatasetShare.java
@@ -12,7 +12,7 @@ public interface UserDatasetShare {
    * The user the dataset is shared with
    * @return
    */
-  Integer getUserId();
+  Long getUserId();
   
   /**
    * The time it was shared
